CalcularTiempoDeEspera con formula de Little

+++ Model/DatoEspera.cs
@@ -6,6 +6,6 @@ namespace Model
     {
         public string Nombre { get; set; } // El nombre de la atracción
         public GoogleMapPosition Posicion { get; set; } // Las coordenadas de la atracción
-        public Dictionary<int,int> TiempoEspera { get; set; } // El tiempo de espera en minutos
+        public Dictionary<int,double> TiempoEspera { get; set; } // El tiempo de espera en minutos
     }
 }
